Section 20 file upload and download file

+++ controllers/admin.js
@@ -2,6 +2,7 @@ const Product = require('../models/product')
 const validationResult = require('express-validator')
 
 const mongoose = require('mongoose')
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
     /*if (!req.session.isLoggedIn) {
@@ -20,9 +21,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title
-    const imageUrl = req.body.imageUrl
+    const image = req.file
     const price = req.body.price
     const description = req.body.description
+    console.log('imageUrl_99: ', image)
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: "Add products", 
+            path:'/admin/edit-product', 
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        })
+    }
+
     const errors = validationResult.validationResult(req)
 
     if (!errors.isEmpty()) {
@@ -42,6 +61,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         })
     }
+
+    const imageUrl = image.path 
 
     const product = new Product({
        // _id: new mongoose.Types.ObjectId('69c02412106436fc5fb374df'),
@@ -119,7 +140,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId
     const updatedTitle = req.body.title
     const updatedPrice = req.body.price
-    const updatedImageUrl = req.body.imageUrl
+    const image = req.file
     const updatedDescription = req.body.description
     const errors = validationResult.validationResult(req)
 
@@ -132,7 +153,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -149,8 +169,13 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle, 
             product.price = updatedPrice, 
-            product.description = updatedDescription,
-            product.imageUrl = updatedImageUrl
+            product.description = updatedDescription
+
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl)
+                product.imageUrl = image.path
+            }
+            
             return product
                     .save()
                     .then(result => {
@@ -187,15 +212,22 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(result => {
-            console.log('Deleted Product!')
-            res.redirect('/admin/products')
-        })
-        .catch(err => {
-            console.error(err)
-            const error = new Error(err)
-            error.httpStatusCode = 500
-            return next(error)
-        })
+    Product.findById(prodId)
+            .then(product => {
+                if (!product) {
+                    return next(new Error('Product not found.'))
+                }
+                fileHelper.deleteFile(product.imageUrl)
+                return  Product.deleteOne({_id: prodId, userId: req.user._id})
+            })
+            .then(result => {
+                console.log('Deleted Product!')
+                res.redirect('/admin/products')
+            })
+            .catch(err => {
+                console.error(err)
+                const error = new Error(err)
+                error.httpStatusCode = 500
+                return next(error)
+            })
 }
